to debug

+++ classes/sphere.cpp
@@ -1,6 +1,6 @@
 #include "sphere.h"
 
-Sphere::Sphere(float x, float y, float z, float r) {
+Sphere::Sphere(int x, int y, int z, int r) {
     this->center_x = x;
     this->center_y = y;
     this->center_z = z;
